move bodo.jit out of top level in tests
