+++ bodo/tests/test_df_lib/test_iceberg.py
@@ -18,12 +18,6 @@ from pyiceberg.table.sorting import SortField, SortOrder
 from pyiceberg.transforms import IdentityTransform
 
 import bodo.pandas as bpd
-from bodo.io.iceberg.catalog.dir import DirCatalog
-from bodo.io.iceberg.catalog.s3_tables import (
-    S3TABLES_REGION,
-    S3TABLES_TABLE_BUCKET_ARN,
-    S3TablesCatalog,
-)
 from bodo.tests.iceberg_database_helpers import pyiceberg_reader
 from bodo.tests.utils import _test_equal
 
@@ -220,6 +214,8 @@ def test_table_read_row_filter(
     table_name,
     memory_leak_check,
 ):
+    from bodo.io.iceberg.catalog.dir import DirCatalog
+
     db_schema, warehouse_loc = iceberg_database(table_name)
     catalog = DirCatalog(
         None,
@@ -265,6 +261,8 @@ def test_table_read_time_travel(
     table_name,
     memory_leak_check,
 ):
+    from bodo.io.iceberg.catalog.dir import DirCatalog
+
     db_schema, warehouse_loc = iceberg_database(table_name)
     catalog = DirCatalog(
         None,
@@ -594,6 +592,12 @@ def test_write():
 
 
 def test_read_s3_tables_location():
+    from bodo.io.iceberg.catalog.s3_tables import (
+        S3TABLES_REGION,
+        S3TABLES_TABLE_BUCKET_ARN,
+        S3TablesCatalog,
+    )
+
     location = "arn:aws:s3tables:us-east-2:427443013497:bucket/tpch"
     region = "us-east-2"
     catalog_properties = {
@@ -613,6 +617,12 @@ def test_read_s3_tables_location():
 
 
 def test_read_s3_tables_read_iceberg_table():
+    from bodo.io.iceberg.catalog.s3_tables import (
+        S3TABLES_REGION,
+        S3TABLES_TABLE_BUCKET_ARN,
+        S3TablesCatalog,
+    )
+
     location = "arn:aws:s3tables:us-east-2:427443013497:bucket/tpch"
     region = "us-east-2"
     catalog_properties = {
@@ -635,6 +645,12 @@ def test_read_s3_tables_read_iceberg_table():
 
 
 def test_write_s3_tables_location():
+    from bodo.io.iceberg.catalog.s3_tables import (
+        S3TABLES_REGION,
+        S3TABLES_TABLE_BUCKET_ARN,
+        S3TablesCatalog,
+    )
+
     location = "arn:aws:s3tables:us-east-2:427443013497:bucket/unittest-bucket"
     region = "us-east-2"
     catalog_properties = {
@@ -671,6 +687,12 @@ def test_write_s3_tables_location():
 
 
 def test_read_iceberg_rename():
+    from bodo.io.iceberg.catalog.s3_tables import (
+        S3TABLES_REGION,
+        S3TABLES_TABLE_BUCKET_ARN,
+        S3TablesCatalog,
+    )
+
     location = "arn:aws:s3tables:us-east-2:427443013497:bucket/tpch"
     region = "us-east-2"
     catalog_properties = {
