+++ bodo/tests/test_spawn/test_recursive.py
@@ -11,35 +11,33 @@ pytestmark = pytest_spawn_mode
 CUSTOMER_TABLE_PATH = datapath_util("tpch-test_data/parquet/customer.pq")
 
 
-@bodo.jit(spawn=True)
-def read_customer_table():
-    df = pd.read_parquet(CUSTOMER_TABLE_PATH)
-    return df
-
-
-@bodo.jit(spawn=True)
-def aggregate_customer_data(df):
-    max_acctbal = (
-        df[df["C_ACCTBAL"] > 1000.0]
-        .groupby(["C_MKTSEGMENT", "C_NATIONKEY"])
-        .agg({"C_ACCTBAL": "max"})
-    )
-    return max_acctbal
-
-
-@bodo.jit(spawn=True)
-def jit_fn_that_calls_other_jit_fns():
-    """Function decorated with @bodo.jit(spawn=True) that calls other functions that are
-    also decorated"""
-    df = read_customer_table()
-    max_acctbal = aggregate_customer_data(df)
-    print("Max acctbal:")
-    print(max_acctbal)
-
-
 def test_recursive():
     """Test that a JIT function that calls another JIT function in spawn mode is
     supported"""
+
+    @bodo.jit(spawn=True)
+    def read_customer_table():
+        df = pd.read_parquet(CUSTOMER_TABLE_PATH)
+        return df
+
+    @bodo.jit(spawn=True)
+    def aggregate_customer_data(df):
+        max_acctbal = (
+            df[df["C_ACCTBAL"] > 1000.0]
+            .groupby(["C_MKTSEGMENT", "C_NATIONKEY"])
+            .agg({"C_ACCTBAL": "max"})
+        )
+        return max_acctbal
+
+    @bodo.jit(spawn=True)
+    def jit_fn_that_calls_other_jit_fns():
+        """Function decorated with @bodo.jit(spawn=True) that calls other functions that are
+        also decorated"""
+        df = read_customer_table()
+        max_acctbal = aggregate_customer_data(df)
+        print("Max acctbal:")
+        print(max_acctbal)
+
     jit_fn_that_calls_other_jit_fns()
