+++ buildscripts/bodo/azure/azure-test-conda-linux-macos.yml
@@ -135,6 +135,12 @@ jobs:
       # For caching tests we need to update the location of the decryption file.
       # We set the absolute path as an environment variable.
       export BODO_TRACING_DECRYPTION_FILE_PATH=`echo "$(System.DefaultWorkingDirectory)/buildscripts/decompress_traces.py"`
+
+      if [[ "${{ parameters.test_type }}" == "SPAWN" || "${{ parameters.test_type }}" == "DF_LIB" ]]; then
+        # Ignore streaming/caching/metadata tests in Dataframes/Spawn mode to avoid importing JIT upfront.
+        PYTEST_IGNORE="--ignore bodo/tests/test_streaming --ignore bodo/tests/caching_tests --ignore bodo/tests/test_metadata.py"
+      fi
+
       if [[ "${{ parameters.test_type }}" == "CACHE" ]]; then
         pixi run -e azure python -m bodo.runtests_caching "${{ parameters.name }}" "$NP" bodo/tests/caching_tests
       elif [[ "${{ parameters.test_type }}" == "SPAWN" ]]; then
@@ -143,11 +149,11 @@ jobs:
         export BODO_ENABLE_DATAFRAME_LIBRARY=0
         export BODO_TEST_SPAWN_MODE=1
         export BODO_NUM_WORKERS="$NP"
-        pixi run -e azure pytest -s -v -Wignore -m "$(PYTEST_MARKER)" bodo/tests --junitxml=pytest-report-spawn-mode.xml --test-run-title="${{ parameters.name }}"
+        pixi run -e azure pytest -s -v -Wignore -m "$(PYTEST_MARKER)" "$(PYTEST_IGNORE)" --junitxml=pytest-report-spawn-mode.xml --test-run-title="${{ parameters.name }}"
       elif [[ "${{ parameters.test_type }}" == "DF_LIB" ]]; then
         export BODO_ENABLE_TEST_DATAFRAME_LIBRARY=1
         export BODO_NUM_WORKERS="$NP"
-        pixi run -e azure pytest -s -v -Wignore -m "$(PYTEST_MARKER)" bodo/tests --junitxml=pytest-report-df-library.xml --test-run-title="${{ parameters.name }}"
+        pixi run -e azure pytest -s -v -Wignore -m "$(PYTEST_MARKER)" "$(PYTEST_IGNORE)" --junitxml=pytest-report-df-library.xml --test-run-title="${{ parameters.name }}"
       else
         # Some tests assume a batch size of 4096 for testing specific behavior.
         export BODO_STREAMING_BATCH_SIZE=4096
